+++ kernel/fs/tmpfs/tmpfs_write.c
@@ -48,7 +48,7 @@ ssize_t tmpfs_write(inode_t* inode, const void* buf, off_t pos, size_t len) {
     tmpfs_inode_t* i = cache_get(&inode->sb->cache, inode->ino);
 
 
-    if(pos + len > i->st.st_size) {
+    if((size_t) pos + len > (size_t) i->st.st_size) {
         
         if(unlikely(
             (
@@ -60,7 +60,7 @@ ssize_t tmpfs_write(inode_t* inode, const void* buf, off_t pos, size_t len) {
         }
 
 
-        if(pos + len > i->capacity) {
+        if((size_t) pos + len > i->capacity) {
 
             i->capacity = pos + len;
             i->capacity = i->capacity + (i->capacity / 2);
